Refactor marginfi_group
