+++ programs/marginfi/fuzz/fuzz_targets/lend.rs
@@ -6,7 +6,7 @@ use arbitrary::Arbitrary;
 use fixed::types::I80F48;
 use lazy_static::lazy_static;
 use libfuzzer_sys::fuzz_target;
-use marginfi::{assert_eq_with_tolerance, prelude::MarginfiGroup, state::marginfi_group::Bank};
+use marginfi::{assert_eq_with_tolerance, prelude::MarginfiGroup, state::bank::Bank};
 use marginfi_fuzz::{
     account_state::AccountsState, arbitrary_helpers::*, metrics::Metrics, MarginfiFuzzContext,
 };
