+++ programs/marginfi/src/instructions/marginfi_account/deposit.rs
@@ -4,8 +4,8 @@ use crate::{
     events::{AccountEventHeader, LendingAccountDepositEvent},
     prelude::*,
     state::{
+        bank::Bank,
         marginfi_account::{BankAccountWrapper, MarginfiAccount, DISABLED_FLAG},
-        marginfi_group::Bank,
     },
     utils::{self, validate_asset_tags},
 };
