+++ programs/marginfi/src/instructions/marginfi_account/repay.rs
@@ -4,8 +4,8 @@ use crate::{
     events::{AccountEventHeader, LendingAccountRepayEvent},
     prelude::{MarginfiError, MarginfiGroup, MarginfiResult},
     state::{
+        bank::Bank,
         marginfi_account::{BankAccountWrapper, MarginfiAccount, DISABLED_FLAG},
-        marginfi_group::Bank,
     },
     utils,
 };
