+++ programs/marginfi/tests/misc/collateral_value_cap.rs
@@ -1,7 +1,7 @@
 use fixtures::{assert_custom_error, prelude::*};
 use marginfi::{
     constants::TOTAL_ASSET_VALUE_INIT_LIMIT_INACTIVE, prelude::MarginfiError,
-    state::marginfi_group::BankConfigOpt,
+    state::bank::BankConfigOpt,
 };
 use pretty_assertions::assert_eq;
 use solana_program_test::*;
