+++ programs/marginfi/tests/misc/operational_state.rs
@@ -2,7 +2,10 @@ use fixed_macro::types::I80F48;
 use fixtures::{assert_custom_error, prelude::*};
 use marginfi::{
     prelude::{GroupConfig, MarginfiError},
-    state::marginfi_group::{BankConfig, BankConfigOpt, BankOperationalState},
+    state::{
+        bank::{BankConfig, BankConfigOpt},
+        marginfi_group::BankOperationalState,
+    },
 };
 use pretty_assertions::assert_eq;
 use solana_program_test::*;
