+++ test-utils/src/bank.rs
@@ -12,7 +12,8 @@ use fixed::types::I80F48;
 use marginfi::{
     bank_authority_seed,
     state::{
-        marginfi_group::{Bank, BankConfigOpt, BankVaultType},
+        bank::{Bank, BankConfigOpt},
+        marginfi_group::BankVaultType,
         price::{OraclePriceFeedAdapter, OraclePriceType, PriceAdapter},
     },
     utils::{find_bank_vault_authority_pda, find_bank_vault_pda},
